Replaced all "long long" types with int64_t Replaced all "unsigned long long" types with uint64_t. These "new" types better communicate their meaning and they are guaranteed to behave the same on all platforms & compilers.

+++ src/libslic3r/GCode/GCodeProcessor.hpp
@@ -7,6 +7,7 @@
 #include "libslic3r/PrintConfig.hpp"
 #include "libslic3r/CustomGCode.hpp"
 
+#include <cstdint>
 #include <array>
 #include <vector>
 #include <string>
@@ -298,7 +299,7 @@ namespace Slic3r {
             PrintEstimatedTimeStatistics time_statistics;
 
 #if ENABLE_GCODE_VIEWER_STATISTICS
-            long long time{ 0 };
+            int64_t time{ 0 };
             void reset()
             {
                 time = 0;
